refactor: cleanup prompt editor
