+++ packages/plugins/plugin-assistant/src/components/TemplateContainer.tsx
@@ -2,92 +2,16 @@
 // Copyright 2025 DXOS.org
 //
 
-import { Match, type Schema as S } from 'effect';
-import React, { type ChangeEvent, useCallback } from 'react';
+import React from 'react';
 
-import { debounce } from '@dxos/async';
-import { Input, Select, Toolbar, useTranslation } from '@dxos/react-ui';
 import { StackItem } from '@dxos/react-ui-stack';
 
 import { TemplateEditor } from './TemplateEditor';
-import { ASSISTANT_PLUGIN } from '../meta';
-import { TemplateKinds, type TemplateKindSchema, type TemplateType } from '../types';
+import { type TemplateType } from '../types';
 
 export const TemplateContainer = ({ template, role }: { template: TemplateType; role: string }) => {
-  const { t } = useTranslation(ASSISTANT_PLUGIN);
-
-  const handleKindChange = useCallback(
-    (value: string) => {
-      const kind = Match.type<string>().pipe(
-        Match.withReturnType<S.Schema.Type<typeof TemplateKindSchema>>(),
-        Match.when('always', () => ({ include: 'always' })),
-        Match.when('schema-matching', () => ({ include: 'schema-matching', typename: '' })),
-        Match.when('automatically', () => ({ include: 'automatically', description: '' })),
-        Match.orElse(() => ({ include: 'manual' })),
-      )(value);
-
-      template.kind = kind;
-    },
-    [template],
-  );
-
-  const handleTypenameChange = useCallback(
-    debounce((event: ChangeEvent<HTMLInputElement>) => {
-      if (template.kind.include === 'schema-matching') {
-        template.kind.typename = event.target.value;
-      }
-    }, 300),
-    [template.kind.include],
-  );
-
-  const handleDescriptionChange = useCallback(
-    debounce((event: ChangeEvent<HTMLInputElement>) => {
-      if (template.kind.include === 'automatically') {
-        template.kind.description = event.target.value;
-      }
-    }, 300),
-    [template.kind.include],
-  );
-
   return (
-    <StackItem.Content role={role} toolbar classNames='container-max-width'>
-      {/* TODO(wittjosiah): Move this toolbar into c11y sidebar. */}
-      <Toolbar.Root>
-        <Select.Root value={template.kind.include} onValueChange={handleKindChange}>
-          <Toolbar.Button asChild>
-            <Select.TriggerButton />
-          </Toolbar.Button>
-          <Select.Portal>
-            <Select.Content>
-              <Select.Viewport>
-                {TemplateKinds.map((kind) => (
-                  <Select.Option key={kind} value={kind}>
-                    {kind}
-                  </Select.Option>
-                ))}
-              </Select.Viewport>
-            </Select.Content>
-          </Select.Portal>
-        </Select.Root>
-        {template.kind.include === 'schema-matching' && (
-          <Input.Root>
-            <Input.TextInput
-              placeholder={t('typename placeholder')}
-              defaultValue={template.kind.typename}
-              onChange={handleTypenameChange}
-            />
-          </Input.Root>
-        )}
-        {template.kind.include === 'automatically' && (
-          <Input.Root>
-            <Input.TextInput
-              placeholder={t('description placeholder')}
-              defaultValue={template.kind.description}
-              onChange={handleDescriptionChange}
-            />
-          </Input.Root>
-        )}
-      </Toolbar.Root>
+    <StackItem.Content role={role} classNames='container-max-width'>
       <TemplateEditor template={template} />
     </StackItem.Content>
   );
