+++ packages/plugins/plugin-assistant/src/translations.ts
@@ -18,7 +18,6 @@ export default [
       },
       [ASSISTANT_PLUGIN]: {
         'templates label': 'Templates',
-
         'open ambient chat label': 'Open AI chat',
         'assistant chat label': 'AI chat',
         'plugin name': 'Assistant',
@@ -29,6 +28,7 @@ export default [
         'command placeholder': 'Enter slash command...',
         'template placeholder': 'Enter template...',
         'value placeholder': 'Enter value...',
+        'prompt rules label': 'Prompt Rules',
         'typename placeholder': 'Enter typename of objects which this template is for',
         'description placeholder': 'Enter description of when this template should be used',
         'select preset template placeholder': 'Select preset',
